# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: HerLoop API
  description: |
    HerLoop 她循环 — 女性社群闲置交易平台后端接口。
    所有接口返回 `{code, message, data}` 格式。分页接口使用 `PageResult<T>` 包装。
  version: 1.0.0
  contact:
    name: HerLoop Team
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: 用户注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
  /auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
  /auth/submit-proof:
    post:
      tags: [Auth]
      summary: 上传实名认证凭证
      description: 上传 Herstory 认证截图，需携带 JWT token。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: integer
                            description: 认证凭证 ID

  # ── User ────────────────────────────────────────────────
...
  /notifications:
    get:
      tags: [Notification]
      summary: 通知列表
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
          description: 仅未读
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSizeDefault20"
      responses:
        "200":
          description: 通知列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageResultNotificationVO"
...
  /notifications/{id}/read:
    put:
      tags: [Notification]
      summary: 标记单条通知已读
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 标记成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
...
  /messages/conversations:
    get:
      tags: [Message]
      summary: 会话列表
      responses:
        "200":
          description: 会话列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ConversationVO"
    post:
      tags: [Message]
      summary: 创建会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetUserId]
              properties:
                targetUserId:
                  type: integer
                productId:
                  type: integer
                  description: 关联商品（可选）
      responses:
        "200":
          description: 会话创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ConversationVO"
  /messages/conversations/{id}/messages:
    get:
      tags: [Message]
      summary: 会话消息列表
      parameters:
        - $ref: "#/components/parameters/idPath"
        - $ref: "#/components/parameters/page"
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: 消息列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageResultMessageVO"
    post:
      tags: [Message]
      summary: 发送消息
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        "200":
          description: 发送成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MessageVO"
  /messages/conversations/{id}/read:
    put:
      tags: [Message]
      summary: 标记会话已读
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 标记成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
...
  /admin/audit/{id}:
    post:
      tags: [Admin]
      summary: 审核认证
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [APPROVED, REJECTED]
      responses:
        "200":
          description: 审核完成
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /admin/users/{id}/ban:
    post:
      tags: [Admin]
      summary: 封禁用户
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 封禁成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /admin/users/{id}/unban:
    post:
      tags: [Admin]
      summary: 解封用户
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 解封成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /admin/add-admin:
    post:
      tags: [Admin]
      summary: 添加管理员
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        "200":
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
...
    pageSizeDefault20:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
...
  schemas:
...
    Result:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          nullable: true

    # ── Auth ──
    RegisterRequest:
      type: object
      required: [email, password, nickname]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        nickname:
          type: string
        avatar:
          type: string
          description: Emoji 头像
        wechat:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResult:
      allOf:
        - $ref: "#/components/schemas/Result"
        - type: object
          properties:
            data:
              type: object
              properties:
                token:
                  type: string
                user:
                  $ref: "#/components/schemas/UserVO"

    # ── User ──
    UserVO:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        avatar:
          type: string
        wechat:
          type: string
        email:
          type: string
        points:
          type: integer
        verified:
          type: boolean
        inviteCode:
          type: string
        productCount:
          type: integer
        tradeCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ── Product ──
...
    SellerBrief:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        avatar:
          type: string
        verified:
          type: boolean

    # ── Want ──
...
    NotificationVO:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        content:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ── Message ──
    ConversationVO:
      type: object
      properties:
        id:
          type: integer
        otherUser:
          $ref: "#/components/schemas/SellerBrief"
        lastMessage:
          type: string
        lastMessageTime:
          type: string
          format: date-time
        unreadCount:
          type: integer
    MessageVO:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Report ──
...
    PageResultNotificationVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/NotificationVO"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    PageResultMessageVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/MessageVO"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer