# herloop-backend/src/main/java/com/herloop/notification/NotificationController.java
package com.herloop.notification;

import com.herloop.common.CurrentUser;
import com.herloop.common.PageResult;
import com.herloop.common.Result;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.*;

import java.util.Map;

@RestController
@RequestMapping("/notifications")
@RequiredArgsConstructor
public class NotificationController {

    private final NotificationService notificationService;

    @GetMapping
    public Result<PageResult<NotificationVO>> list(
            @RequestParam(required = false) Boolean unreadOnly,
            @RequestParam(defaultValue = "1") Integer page,
            @RequestParam(defaultValue = "20") Integer pageSize) {
        Long userId = CurrentUser.getId();
        return Result.success(notificationService.list(userId, unreadOnly, page, pageSize));
    }

    @GetMapping("/unread-count")
    public Result<Map<String, Long>> unreadCount() {
        Long userId = CurrentUser.getId();
        return Result.success(Map.of("count", notificationService.getUnreadCount(userId)));
    }

    @PutMapping("/{id}/read")
    public Result<Void> markAsRead(@PathVariable Long id) {
        Long userId = CurrentUser.getId();
        notificationService.markAsRead(userId, id);
        return Result.success(null);
    }

    @PutMapping("/read-all")
    public Result<Void> markAllAsRead() {
        Long userId = CurrentUser.getId();
        notificationService.markAllAsRead(userId);
        return Result.success(null);
    }
}